refactor(pipeline): extract pause functions to pause.go [#466]
Move drainPauseSignal, waitIfPaused, and Confirm from engine.go to pause.go.
