+++ internal/pipeline/engine.go
@@ -156,70 +156,6 @@ func NewEngine(r runner.Runner, state *State, cfg EngineConfig) *Engine {
 	return e
 }
 
-// drainPauseSignal reads from the pause channel and updates the paused flag.
-// When a resume signal (false) arrives while the engine is blocked on a
-// checkpoint (inCheckpoint), the method also sends to confirmCh to unblock
-// the checkpoint wait — without this, the engine deadlocks because it waits
-// on confirmCh, not pauseCond.
-// When the channel is closed (TUI exits), the goroutine force-unpauses to
-// unblock any waiters, preventing deadlock.
-func (e *Engine) drainPauseSignal(ch <-chan bool) {
-	for p := range ch {
-		e.pauseMu.Lock()
-		e.paused = p
-		if !p {
-			e.pauseCond.Broadcast()
-			// If the engine is blocked on a checkpoint, unblock it.
-			if e.inCheckpoint && e.confirmCh != nil {
-				select {
-				case e.confirmCh <- struct{}{}:
-				default:
-				}
-			}
-		}
-		e.pauseMu.Unlock()
-	}
-	// Channel closed: force-unpause to unblock any waiters.
-	e.pauseMu.Lock()
-	e.paused = false
-	e.pauseCond.Broadcast()
-	// Also unblock checkpoint if blocked.
-	if e.inCheckpoint && e.confirmCh != nil {
-		select {
-		case e.confirmCh <- struct{}{}:
-		default:
-		}
-	}
-	e.pauseMu.Unlock()
-}
-
-// waitIfPaused blocks until the engine is unpaused or context is cancelled.
-// Returns ctx.Err() if context was cancelled while paused, nil otherwise.
-func (e *Engine) waitIfPaused(ctx context.Context) error {
-	e.pauseMu.Lock()
-	defer e.pauseMu.Unlock()
-	for e.paused {
-		// Check context before waiting
-		if err := ctx.Err(); err != nil {
-			return err
-		}
-		// Use a goroutine to wake on context cancellation
-		done := make(chan struct{})
-		go func() {
-			select {
-			case <-ctx.Done():
-				e.pauseMu.Lock()
-				e.pauseCond.Broadcast()
-				e.pauseMu.Unlock()
-			case <-done:
-			}
-		}()
-		e.pauseCond.Wait()
-		close(done)
-	}
-	return ctx.Err()
-}
-
 // checkBinaryVersion records the binary version in meta on first run and
 // emits a warning event if the current binary differs from the version
 // that originally created the pipeline state.
@@ -585,13 +521,6 @@ func (e *Engine) executePhases(ctx context.Context, phases []PhaseConfig, forceF
 		if !rework {
 			return nil
 		}
-	}
-}
-
-// Confirm unblocks the engine in Checkpoint mode.
-func (e *Engine) Confirm() {
-	if e.confirmCh != nil {
-		e.confirmCh <- struct{}{}
 	}
 }
